Added the ability to ignore specific packages
You can do this by putting white space separated
packages in the packages.ignored file in the
master branch
